단축한 URL을 통해 사이트 이동이 가능하도록 변경
