+++ requirements.txt
@@ -1,22 +1,40 @@
+aiohttp
 appdirs
 asgiref
+async-timeout
+attrs
 black
+certifi
+chardet
+charset-normalizer
 click
 Django
 django-debug-toolbar
 django-ninja
+django-ratelimit
 django-seed
+django-user-agents
 djangorestframework
 Faker
+geoip2
 gunicorn
+idna
+maxminddb
+multidict
 mypy-extensions
+mysqlclient
 pathspec
 pydantic
 python-dateutil
 pytz
 regex
+requests
 six
-sqlparse
+sqlparse=
 text-unidecode
 toml
 typing-extensions
+ua-parser
+urllib3
+user-agents
+yarl
