+++ shortener/models.py
@@ -79,3 +79,13 @@ class ShortenedUrls(TimeStampedModel):
     create_via = models.CharField(
         max_length=8, choices=UrlCreatedVia.choices, default=UrlCreatedVia.WEBSITE)
     expired_at = models.DateTimeField(null=True)
+
+    class Meta:
+        indexes = [
+            models.Index(
+                fields=[
+                    "prefix",
+                    "shortened_url",
+                ]
+            ),
+        ]
