+++ shortener/urls/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 from shortener.forms import UrlCreateForm
 from shortener.models import ShortenedUrls
 from django.contrib import messages
@@ -76,3 +76,18 @@ def url_change(request, action, url_id):
         return render(request, "url_create.html", context)
 
     return redirect("url_list")
+
+
+def url_redirect(request, prefix, url):
+    print(prefix, url)
+    get_url = get_object_or_404(
+        ShortenedUrls, prefix=prefix, shortened_url=url)
+    is_permanent = False
+    target = get_url.target_url
+    if get_url.creator.organization:
+        is_permanent = True
+
+    if not target.startswith("https://") and not target.startswith("http://"):
+        target = "https://" + get_url.target_url
+
+    return redirect(target, permanent=is_permanent)
